[RATOM-241] Adds a new file status

+++ etl/tasks.py
@@ -27,9 +27,13 @@ def remove_file_task(files: []):
     has no files, after an attempted restore.
     """
     for f in File.objects.filter(pk__in=files):
+        if f.import_status != File.FAILED:
+            continue
         logger.info(
             f"Recovering {f.account.title} by removing failed file: {f.filename}"
         )
+        f.import_status = File.RESTORING
+        f.save()
         f.delete()
         if f.account.files.count() == 0:
             f.account.delete()
